requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.4.0
python-multipart==0.0.22
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator>=2.0.0